Add files via upload
In this project we intend to use a dataset obtained from a thorough chemical examination of these wines in order to use machine learning to identify wines from various varietals grown in the same region of Italy.

There are 178 examples in the collection, each of which represents a unique wine sample with 13 chemical components, such as alcohol percentage, acidity, phenolic compounds, and color intensity. The wine cultivar, which is divided into three different groups, is the target variable for this classification task. 

 We will use Random Forest, k-Nearest Neighbors (k-NN), Support Vector Machines (SVM), and Logistic Regression to test the dataset. We want to determine the best method for precisely forecasting the wine cultivar based on its chemical characteristics by creating a baseline model and contrasting the results of various algorithms.